pyproject.toml
[project]
name = "omotemmywallet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt[crypto]>=5.5.1",
    "djangorestframework-stubs>=3.16.8",
    "pip>=26.0.1",
    "psycopg2-binary>=2.9.11",
]
